fix: some windows warnings

+++ src/Features/AutoSubmitMod.cpp
@@ -398,7 +398,7 @@ void AutoSubmitMod::FinishRun(float final_time, const char *demopath, std::optio
 
 	auto map_id = it->second;
 
-	int score = floor(final_time * 100);
+	int score = (int)floor(final_time * 100);
 
 	if (g_worker.joinable()) g_worker.join();
 	g_worker = std::thread(submitTime, score, std::string(demopath), Utils::StartsWith(engine->GetCurrentMapName().c_str(), "mp_"), map_id, rename_if_pb, replay_append_if_pb);
